core/coroutine: add a specialization for std::expected<void, Err>
As recorded in https://www.scylladb.com/2024/05/14/better-goodput-performance-through-c-exception-handling/
under `Approach 1: Avoid The Exceptions` there are places where we would
like to avoid exceptions. One approach we do is use `boost::result`, but
now that we're on C++23 have instead favored using `std::expected` as
it's a bit simpler and "modern" (for what that is worth).

However, for side effecting functions we've started to represent that
with `std::expected<void, Err>`, which is fine, but very cumbersome to
fully type out. We'd like to specialize coroutine_traits to make this case
less verbose, but can't do it in our application due to it being in an
`internal` namespace and compatibility not being ensured.

So we propose a specialization for `std::expected` that allows the
syntax `co_return {};`. This is really a small quality of life
improvement for us we'd like to understand if there is an appetite for
having upstream.