coroutine: introduce try_future
Special awaiter which co_await:s a future and returns the wrapped result
if successful, terminates the coroutine otherwise, propagating the
exception directly to its waiter.

If the future was successful, this is identical to co_await-ing the future
directly. If the future failed, the coroutine is not resumed and instead the
exception from the future is forwarded to the waiter directly and the
coroutine is destroyed.

The goal of this special awaiter is to provide the good ergonomics of
throw -- interrupting control flow and immediately propagating the
exception to the caler -- without the associated costs. The exception is
propagated via the future chain, without being thrown.