coroutine: introduce try_future
Special awaiter which co_await:s a future and returns the wrapped result
if successful, terminates the coroutine otherwise, propagating the
exception directly to its waiter.

If the future was successful, this is identical to co_await-ing the future
directly. If the future failed, the coroutine is not resumed and instead the
exception from the future is forwarded to the waiter directly and the
coroutine is destroyed.

The goal of this special awaiter is to provide the good ergonomics of
throw -- interrupting control flow and immediately propagating the
exception to the caler -- without the associated costs. The exception is
propagated via the future chain, without being thrown.

The new awaiter uses
execute_involving_handle_destruction_in_await_suspend(), but it has to
massage it a bit first: the awaiter is already a task itself, so to
prevent unnecessary allocation of lambda task,
execute_involving_handle_destruction_in_await_suspend() now checks if
func is a seastar task and will schedule it directly in this case.

The coroutine exception handling section in doc/tutorial.md is updated
to also mention coroutine::try_future (and while there
coroutine::as_future as well).