rpc: sink_impl: batch sending and deletion of snd_buf:s
Define batched_queue that is used for batch
processing of snd_buf:s on a remote shard.

It id used first to queue up buffers on the send path
where a send loop is invoked on the connection shard
to send queued batches of snd_buf:s.

Then, on the completion path, the exhausted buffers
are queued for deletion on the delete queue, where
the processing loop is invoked back on the sink shard
to delete the buffers.

Both changes avoid too long task queues that may be caused
by sending small messages across shards.

Note that batched_queue ensures processing of the
buffers in fifo order also across shards, so the equence_number
mechanism previously used to reorder out-of-order continuations
was dropped.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>