io: Count and export number of AIO retries
The AIO requests may come via two places -- the reactor itself polling
for IO and thread-pool that retries requests that had been rejected by
the kernel on the first try.

This patch counts the latter events and exports them as metrics. The
former can be calculated using existing aio-reads and aio-writes.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>