io_tester: Don't skip 0 position in sequential workflows
When performing IO sequentially, io_tester's get_pos() helper adds
request size to initial value of the "current position" and then returns
it, thus skipping the 0th position in a file.

Fix it, and rename the _last_pos to be _next_seq_pos to better reflect
what the field is about.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>