http: Explicitly deprecate request::content
Modern httpd handlers must use request body stream to get the contents
of the request, but this "must" is in fact "should" as there's no such
visible requirement other than a code comment nobody reads.

Http client callers are not supposed to access this member either,
there's write_body() method for that. But this time it's simpler -- the
client is still in the experimental namespace.

Still we need to handle httpd users and deprecate content explicitly.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>