io_queue: Destroy priority class data with scheduling group
When a scheduling group is destroyed, the IO queue's priority_class_data
that corresponds to it stays intact. It's not actually a leak, as when
the reactor is destroyed all class data objects are collected. And even
if another scheduling group under the same name is created, the IO queue
will re-use the dangling priority class for the newly created group.

However, the fact that the prioroty class stays alive is not nice. This
patch implements priority class destruction. For that, several things
are to be handled.

First, priority class data may have queued requests. That's fairly
simple, all those requests can be canceled and disposed.

Next, priority class data may have executing requests. That's worse, as
they cannot be canceled. The destroying code could wait for them to
complete, but since it's likely going to be rare condition, it seems
simpler to let the class dissolve in the background.

Finally, the respective fair-queue classes has to be unregistered too,
but that's fairly simple.

Test included.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>