execution_stage: move metrics setup to a separate method
The purpose of this change is to allow reuse of the code in the next
commit of this patch series.

This commit also adds a `_metric_group.clear()` call before assigning
new metric_group, to make sure we don't register the same statistics
twice when update_metric_group() is called multiple times.

Refs: https://github.com/scylladb/seastar/issues/2992