condition_variable: support aborting condition variable waits
Sometimes you have an abort source, and the trick to signal the
condition variable to wake up your waiter is not ideal as there might be
many waiters and you don't want to wastefully invoke all of their
predicates. Instead push the abort source into the condition variable so
we can remove the waiter directly from the intrusive list.