future: handle lvalue references in future continuations early
Typically, future continuations are passed as temporaries, since they
are usually lambdas written at the point of call:

    f.then([] { ... });

Rarely, an lvalue is passed:

    auto const = [] { ... };
    f.then(cont);

To avoid having to track the lvalue-vs-rvalue status of the continuation
everywhere in the complicated continuation machinery, convert
lvalues to rvalues early with std::ref(). The resulting
std::reference_wrapper object will be a temporary, but will behave as a
reference (including in its memory storage costs).

This simplifies further changes in this area.