future: invoke continuation functions as temporaries
Continuations are almost, but not quite, always invoked as temporaries:

    f.then([] { ... });

The lambda passed to future::then() is a temporary.

Rarely, lvalues are passed as continuations:

    auto cont = [] { ... };
    f.then(cont);

This isn't super useful but exists in the wild.

Previously, the lvalue-vs-rvalue-ness of the continuation function
wasn't well tracked. But the previous patch converted *all*
continuation functions to rvalues, even if they were originally
passed as lvalues (using std::ref()). This means that anything
downstream of future entry points like future::then() can treat
the continuation function type as a non-reference.

We can therefore invoke the continuation as a temporary.

This rarely makes any difference, since it is rare to define
the function call operator as taking its object as a temporary (e.g.
`void operator() &&`) but will become common when we use "deduced this"
(`void operator(this auto) { ... }` or more commonly its lambda form).

To accomodate rvalue-qualified continuations, change the continuation
machinery to invoke continuations as temporaries.