coroutines: advertise lambda-capture-by-value and test it
Since lambdas have the signature `ret_type lambda::operator()(...) const &`,
the lambda itself is not copied to the coroutine frame. This is okay
if the C++ lifetime extension rules extend the lambda's lifetime
(if the coroutine is awaited in the same statement), but not okay
if the coroutine's lifetime exceeds the current statement's.

Luckily, there is a way to convert the signature to accept the lambda
by value: `ret_type lambda::operator()(this auto, ...)`. The hidden
object parameter is now passed by value, and if this is a lambda
coroutine, it will be copied into the coroutine frame, divorcing its
lifetime from that of its creator.

Document this happy finding and add tests to prove that it works.