core: restore seastar_logger namespace in try_systemwide_memory_barrier
`try_systemwide_memory_barrier` was recently moved to the
`seastar::internal` namespace in cf933e7abdbb444143a1c13942dd4ff7e134f800.
However, the function contained an `extern logger seastar_logger`
declaration for `__aarch64__`. As a result, the logger symbol was
treated as if it were declared inside `seastar::internal`, which led to
a link error on ARM.

Fixes: https://github.com/scylladb/seastar/issues/3021

Closes scylladb/seastar#3022