io_tester: add aio fallbacks count
Add aio fallbacks metrics (the io_threaded_fallback metrics with
"reason" aio_fallback), as this is an important aspect of IO performance
which is getting more investigation at the moment.

Note that this is the count of fallback "batches" submitted to the
syscall thread, not the number of individual IO ops that fell back (as
one batch may have more than one IO in it).