output_stream: Unify some code-paths of slow_write()
The function is called when the passed memory buffer cannot be just
appended to the existing stream buffer and it falls into three
sub-paths.

First, when there's some data buffered (and the passed buffer doesn't
fit the remaining tailroom) -- the head of the passed buffer is
appended, the inner buffer is put into sink, the remainder of the passed
buffer becomes the inner buffer.

Second, when there's some data buffered, but the passed buffer is too
large for the first step to happen. In that case the passed buffer is
also split and its head is appended to the inner buffer, but then both
inner buffer and remainder of the passed one are put into sink.

Third case is when there's no data buffered. In that case the passed
buffer is just put into sink as is.

It's seen that first two cases are very similar. However, looking at the
slow_write() doesn't make it obvious, instead those cases are
implemented independently, and rely on obscure bulk_threshold variable
calculation. Merging those two paths hopefully helps.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>