rpc: Wrap socket and buffers
One of the connection class initialization is performed "lazily" --
first the class is constructed and its socket and in/out buffers are
default-initialized. Then the creator connects for real and calls
set_socket() method to setup socket and buffers. And the bool _connected
bit is also set indicating that the connection is ready.

This patch packs socket and buffers into optional struct and replaces
the _connected bool with optional::has_value() checks.

The change is needed, because streams' default constructor and
move-assignment operator are going to be deprecated.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>