iostream: Deprecate input/output stream default constructor and move-assignment operator
The patter when stream is default-constructed in one place and is then
move-assign overwritten in another is pretty nasty.

Current implementation of operator= is incorrect, it just overwrites the
objects, but it ideally it should "abort" the stream if it's alive.
Because of that, the stream should better not be used in between, and
all the users of that pattern try to do it, but it's rather fragile.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>