resource: don't error if --overprovisioned and asking for more vcpus than available
The --overprovisioned parameter (which should really be named --overcommit)
says we're okay with overloading the machine. Often because it's a
developer machine shared with a development environment, but could also
be just for testing. Allow starting more shards than there are vcpus
in that case. This is useful for testing large shard counts.