memory: make cpu_id_shift and related mask dynamic
Right now cpu_id_shift allows for 8 bits of cpu_id so a maximum
of 256 shards.

Relax that and make it dynamic, calculating it in a new function
global_setup().

We leave it at 8 bits even when the shard number can be stored in
fewer bits, so we retain the traditional memory map
0xsccv'vvvv'vvvv'vvvv. If more bits are needed they will be taken
from the most significant bits of v.

Note: cpu_id_shift and cpu_id_and_mem_base_mask can be used
even before global_setup is called, due to allocations that
predate the calculation of original_malloc_func, so they are
initialized to reasonable values. Overriding them later will retain
memory map compatibility.