iotune: fix very long warm up duration on systems with high cpu count
In an i7i.16xlarge instance iotune with default duration of 120s
would run for 120s + 18s + 48 * 6s = 426s which blows most of
our timeouts in Scylla setup scripts.

This patches changes the warmup period calculation so that for the
sequential write step, it stays a total of the original
max(3s,5% * duration) for all shards.

Signed-off-by: Robert Bindar <robert.bindar@scylladb.com>