test: Add unit test for append_challenged_posix_file
The test validates several things:
- the maximum number of writes-above-file-size doesn't exceed the limit
- the number of writes-above-file-size actually reaches it
- no parallel fsyncs are run if disabled

In fact, the 2nd test fails for concurrency above 1, but for XFS it's
not currently configured, so OK for now.

To work, the test creates an "real" temporary file, but doesn't do real
IO, instead, it runs counting-only operations. Truncations, however,
happen for real, because this file impl uses raw ftruncate systemcall
for it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>