http: Switch content_length_source_impl to update remaining bytes
This source currently can be asked to track the total number of consumed
bytes. This is for the client to later compare this number with the
response content_length and decide how to coninue with the connection.

This patch changes this math to use only remaining bytes left for the
source to read. It makes the source_impl simpler, the client deciding
code simpler and prepares the ground for the chunked-encoded source to
do the same check.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>