http: Track "remaining bytes" for chunked_source_impl
Client uses this number to decide what to do with the connection after
reply handler callback finihsed working with the response.

For content-legnth encoded body, if the handler leaves some unread bytes
(the remaining bytes counter is non-zero) the client either skips those
bytes from the socket (by reading the necessary amount of bytes and
instantly discarding them), if that number is low, os marks the
connection for closing.

For chunked-encoded body the very same check cannot be done, because the
size of the body is not known up until the whole body is read. So to
detect the case when handler leaves the reply body unread is done in
yes/no manner -- the source sets the "remaining bytes" to non-zero value
and only resets it to zero after all response is read.

Similarly, to prevent the client from skipping the remaining bytes from
the conneciton, its initial value is set to be infinite to be greater
than any threshold that can be configured on the client.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>