metric:value: implement + in terms of +=
We implemented += in terms of +, but it is almost always more efficient
to do the reverse, as + creates a new object (which cannot easily be
optimized away in the += case), whereas += does not. Furthermore, we
heavily use += during metrics aggregation, whereas we don't use + in
the seastsar codebase at all.

This is good for a > 5% improvement in aggregation heavy prometheus
benchmarks.