memory.cc: compile fix for std::bit_width()
After recent changes to memory.cc, it fails to compile as standard
libraries vary in their return value for std::bit_width. Initial C++20
behavior was to return an integer of the same signeness as the input,
but later standards specify an `int` return type.

This patch changes the variable `w` to be an `int` from auto so it
works everywhere.

Closes scylladb/seastar#3067