smp: Fix exception safety of invoke_on_... internal copying
Both invoke_on_all and invoke_on_others copy the function passed to it
to be then submit_to()-ed on other shards. And only the invoke_on_others
requires that the function is nothrow-copy-constructible. Marking the
invoke_on_all() to requires nothros-copy-constructible function is
hardly possible, there's lots of code out there that already passes
functions with captured strings that can throw when being copied. So
this change checks the if function is nothrow copyable and just copies
that, otherwise it captures the potential exception and returns it as a
future. The invoke_on_others is patched to do the same, for symmetry,
and its requirement to have nothrow copyable function is relaxed.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>