net: tighten port load balancing port access
We fetch the port from the sockaddr structure as if it's always
sockaddr_in, where it could also be sockaddr_in6.

This doesn't cause any problems, since in both sockaddr_in and
sockaddr_in6 the port is in the same offset. However, it's not nice
to access using the wrong type and to rely on such accidents.

"Fix" by dispatching based on the actual family.

A unit test is added to make sure we don't regress. It passes
both before and after the patch.