test: Make output_stream_test splitting test case use own sink
The test validates that pre-defined set of strings is re-split by the
output_stream bufferring into some other set of strings. For that it
accumulates the produced buffers in a packet and then decomposes it into
fragemtns to compare each with the expected result vector.

This patch avoids using intermediate packet and compares the produced
buffers with expected "on the fly".

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>