net/posix-stack: Don't use packet in posix_data_sink_impl
The place to patch is ::put(span<temporary_buffer>) method in 9th API
level. When called the method converts span of buffers into packet and
pushes it into pollable_fd::write_all(). This patch makes the put()
method convert span of buffers into a vector of iovec-s and push the
reference to the vector into pollable_fd::write_all(). The pollable_fd,
in turn, transparently converts the vector of iovecs into span of iovecs
to call reactor::sendmsg().

The pollable_fd::write_all(net::packet) overload is kept for the 8th API
level.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>