timer: Enable highres timer based on next timeout value
When highres timers are completed, the post-completion callback checks
if there are any timers queued to decide whether or not to set up the
timerfd notification. The check is wrong, timer list can be empty at that
point if a timer is (was) canceled while being run. At the same time,
since "next timeout" value is set, timer-set thinks that the timerfd is
active and may decide not to re-activate it when another timer appears.

The fix is to engage timerfd based on "next timeout" value.

Test that reproduced the issue is included, it fails before the fix.

fixes #3013

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>

Closes scylladb/seastar#3092