github: disable AVX2 instruction set when building dpdk
by default, we build dpdk with "meson setup
-Dcpu_instruction_set=native". in which, "cpu_instruction_set=native"
option tells the building system to configures "-march", "-mcpu" and
"-mtune" to "native". this might instruct the compiler to emit target
code using advanced instruction set like AVX-512, while the the CPU
might not be able to execute it, because the hypervisor might not
expose all CPU features in a virtualization environment.

in this change, instead of using "generic" which is mapped to "corei7",
we use a conservative but performant alternative -- "corei7-avx"
where SSE4.2 and AVX is enabled, but AVX2 is disabled.

Fixes #3083
Signed-off-by: Kefu Chai <k.chai@proxmox.com>

Closes scylladb/seastar#3093