rpc: Add a new constructor in closed_error to accept string argument
In the scylladb codebase, it has been observed that the presence of certain context
details in the error messages can ease debugging in case of error throws. For such
scenarios, its helpful to have an interface in the error class itself that accepts those
information and passes them along to enrich the context.

This commit adds a new constructor in the rpc::closed_error which accepts a string as
argument, where the caller can pass relevant context useful for debugging.

Since this is not a bug fix, backport is not needed

ref: scylladb/scylladb#16923