posix-stack: add assert to put(packet)
Previously we had put an assert on the usage of the packet in reactor
but this was too expensive for our perf tests. The next best thing is to
police the on-ramps to where the null ptr error occurred.

Adds a check that when a packet is put to an iostream, any prior data to
be written has already been cleared by the completion of the write_all
continuation