reactor: remove blocking of SIGILL
It's not clear why we do this, while the application shard (smp::_tmain) has a bunch of signals unblocked on it. This is annoying in Redpanda moreso now because we have turned on libc++ hardening mode[1] in our debug build, which crashes with SIGILL, and currently we don't get signals for that.

Partially Fixes: https://github.com/scylladb/scylladb/issues/14163

[1]: https://libcxx.llvm.org/Hardening.html