coroutine: exception: workaround broken destroy coroutine handle in await_suspend
gcc 15.2 and above crash if a coroutine handle is destroyed when
await_suspend for that coroutine is called [1]. An internal reference
counter is accessed after the coroutine is destroyed.

Work around this by scheduling the work to a task. This is slower,
but doesn't crash, and exceptions are supposed to be somewhat rare.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=121961