scripts: add 'integer' type to seastar-json2code
Swagger 2.0 expects the integer type to be named `integer`, whereas
`seastar-json2code` only supports the `int` type. As a result, running
`seastar-json2code` on a model that contains the `integer` type fails.

This commit extends the list of supported types to allow `integer`
in addition to `int`. The old naming (`int`) is retained for backward
compatibility.

Fixes: scylladb/seastar#3103