reactor: Update loads evaluation loop
The loop maintains 5 recent "load" values. Every tick it substracts the
oldest value from the accumulator and adds the newest one. For that
circular buffer is used and there are two minor inefficiencies with it

1. the buffer is grown on demand and grows up to 8 points, while on 5
   are in use
2. the buffer size is checked each tick if it has grown large enough,
   and most of the time the answer is -- yet it has

This patch reserves the buffer to 5 records in advance and pre-populates
it with zeroes so that check in timer is not needed.

Also add named constant for the number of points "while at it".

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>