dns: Replace net::packet with vector<temporary_buffer>
Sending the packet via tcp happens with the help of data_sink::put()
which have an overload that accepts vector of buffers.

Sending the packet via udp still happens with the help of intermediate
packet construction, but it will be fixed eventually by converting
datagram_channel-s to use span<temporary_buffer> instead of packet-s.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>