dns: Optimize packet sending for newer c-ares versions
Recently newer c-ares support was added. In it, the do_sendv is always
called with a single buffer entry. For that case, there's no need to
construct a vector with a single element. Using standalone buffer is
shorter and renders less allocations.

The tcp-sending compiles naturally with the new type of sent packet, for
udp-sending there's an ifdef to create single-element span.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>