memory: add C23 free_sized, free_aligned_sized
Similar to C++ sized deallocation, but for the C malloc family. They are
in glibc 2.43, not yet released.

The implementation for our own memory allocator is trivial, we just
forward the call to the allocator.

For debug mode, it's a little tricky - we typically don't override
standard library functions when using the system allocator, but these
functions won't be available for a while. We define them conditionally
here.

The need comes from class replacements for operator new/delete. If we
have them forward to global operator new/delete:

```c++
struct foo {
    void* operator new(size_t sz) { return ::operator new(sz); }
    void* operator delete(void* p, size_t sz) { return ::operator delete(p, sz); }
}
```

Then gcc is clever enough to figure out the the eventual deletion is done
by the global operator delete, but not clever enough to figure out that the
allocation was done by the global operator new, and throws a fit over
the mismatched allocation/deallocation functions.

The solution is to use malloc/free, but we don't have a matching sized
free function, and don't want to lose the optimization. So here we add
a sized free that works like sized operator delete.