reactor,file: Get and populate fs_info in opening thread pool
Right now the fs_info map is populated in the reactor thread. This is
done it make_file_impl -- it calls reactor::fstatfs(), that switches to
thread-poll, then it instantiates an fs_info object in global map.

After this patch, calling fstatfs and populating the map of fs_info-s is
done in the thread pool. The caller in this case is reactor::open_file_dma
and reactor::open_directory that do opening in thread-pool anyway.

The returned pointer to fs_info object is then passed (in reactor thread)
to make_file_impl() which no longer needs to access the map of fs_infos.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>