code: Move pollable_fd_state::write_all(const char*) from API level 9
The pair of methods sending plain buffer into fd are not used in the
9th API level, because the only caller of it are data_sink::put()-s
overloads that send single temporary_buffer and they were left in older
API levels in favor of span-nized put()-s.

Together with this method, reactor::send() bunch. All those methods but
the send_all() one are private. The send_all, however, needs internal
pollable_fd_state& argument, so is likely not the part of the public API
either.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>